requirements.txt
streamlit
openai
python-dotenv
pandas
python-decouple
requests
